feat(tray): Implement watch threshold settings for Linux and macOS
- Added a threshold submenu in the system tray for Linux using AppIndicator.
- Implemented custom threshold input using zenity for Linux.
- Added macOS-specific implementation for custom threshold input using AppleScript.
- Created a new Windows tray application with similar functionality, including custom threshold input using Tkinter.
- Refactored icon loading methods to utilize base class functionality across platforms.
- Enhanced error handling and logging for better debugging and user notifications.
