+++ simkl_mps/main.py
@@ -18,11 +18,11 @@ from simkl_mps.credentials import get_credentials
 def get_tray_app():
     """Get the correct tray app implementation based on platform"""
     if sys.platform == 'win32':
-        from simkl_mps.tray_app import TrayApp, run_tray_app
+        from simkl_mps.tray_win import TrayAppWin as TrayApp, run_tray_app
     elif sys.platform == 'darwin':
-        from simkl_mps.tray_mac import TrayApp, run_tray_app
+        from simkl_mps.tray_mac import TrayAppMac as TrayApp, run_tray_app
     else:  # Linux and other platforms
-        from simkl_mps.tray_linux import TrayApp, run_tray_app
+        from simkl_mps.tray_linux import TrayAppLinux as TrayApp, run_tray_app
     return TrayApp, run_tray_app
 
 class ConfigurationError(Exception):
